lecture-11-Capstone-project/requirements.txt

streamlit>=1.30.0
openai>=1.3.5
pandas>=2.0.0
python-dotenv>=1.0.0
